some minor comments fix
