+++ AmplifyPlugins/Predictions/AWSPredictionsPlugin/Support/Utils/IdentifyResultTransformers.swift
@@ -47,7 +47,6 @@ class IdentifyResultTransformers {
             points.append(point)
         }
         return Polygon(points: points)
-
     }
 
     static func processPolygon(_ textractPolygonPoints: [AWSTextractPoint]?) -> Polygon? {
@@ -64,90 +63,6 @@ class IdentifyResultTransformers {
             points.append(point)
         }
         return Polygon(points: points)
-
-    }
-
-    static func processChildOfKeyValueSet(ids: [String],
-                                          blockMap: [String: AWSTextractBlock]) -> String {
-        var keyText = ""
-        for keyId in ids {
-            guard let keyBlock = blockMap[keyId],
-                let text = keyBlock.text,
-                case .word = keyBlock.blockType else {
-                continue
-            }
-            keyText += text + " "
-        }
-        return keyText.trimmingCharacters(in: .whitespacesAndNewlines)
-    }
-
-    static func processValueOfKeyValueSet(ids: [String],
-                                          blockMap: [String: AWSTextractBlock]) -> (String, Bool) {
-        var valueText = ""
-        var valueSelected = false
-
-        for valueId in ids {
-            let valueBlock = blockMap[valueId]
-            guard let valueBlockRelations = valueBlock?.relationships else {
-                continue
-            }
-            for valueBlockRelation in valueBlockRelations {
-                guard let wordBlockIds = valueBlockRelation.ids else {
-                    break
-                }
-                for wordBlockId in wordBlockIds {
-                    let wordBlock = blockMap[wordBlockId]
-                    guard let wordValueBlockType = wordBlock?.blockType else {
-                        continue
-                    }
-                    switch wordValueBlockType {
-                    case .word:
-                        if let text = wordBlock?.text {
-                            valueText += text + " "
-                        }
-                    case .selectionElement:
-                        valueSelected = wordBlock?.selectionStatus == .selected
-                    default: break
-                    }
-                }
-            }
-        }
-        return (valueText.trimmingCharacters(in: .whitespacesAndNewlines), valueSelected)
-    }
-
-    static func processLineBlock(block: AWSTextractBlock) -> IdentifiedLine? {
-        guard let text = block.text,
-            let boundingBox = processBoundingBox(block.geometry?.boundingBox),
-            let polygon = processPolygon(block.geometry?.polygon) else {
-                return nil
-        }
-
-        return IdentifiedLine(text: text,
-                              boundingBox: boundingBox,
-                              polygon: polygon,
-                              page: Int(truncating: block.page ?? 0))
-    }
-
-    static func processWordBlock(block: AWSTextractBlock) -> IdentifiedWord? {
-        guard let text = block.text,
-            let boundingBox = processBoundingBox(block.geometry?.boundingBox),
-            let polygon = processPolygon(block.geometry?.polygon) else {
-                return nil
-        }
-
-         return IdentifiedWord(text: text,
-                               boundingBox: boundingBox,
-                               polygon: polygon,
-                               page: Int(truncating: block.page ?? 0))
-    }
-
-    static func processSelectionElementBlock(block: AWSTextractBlock) -> Selection? {
-        guard let boundingBox = processBoundingBox(block.geometry?.boundingBox),
-            let polygon = processPolygon(block.geometry?.polygon) else {
-                return nil
-        }
-        let selectionStatus = block.selectionStatus == .selected ? true : false
-        return Selection(boundingBox: boundingBox, polygon: polygon, isSelected: selectionStatus)
     }
 
     // swiftlint:disable cyclomatic_complexity
